testAdd.sh

+++ calculator.c
@@ -1,21 +1,21 @@
 #include "calculator.h"
 
-int add(int getal1, int getal2)
+float add(int getal1, int getal2)
 {
 	return getal1 + getal2;
 }
 
-int sub(int getal1, int getal2)
+float sub(int getal1, int getal2)
 {
 	return getal1 - getal2;
 }
 
-int mul(int getal1, int getal2)
+float mul(int getal1, int getal2)
 {
 	return getal1 * getal2;
 }
 
-int div(int getal1, int getal2)
+float div(int getal1, int getal2)
 {
 	return getal1 / getal2;
 }
